+++ src/main/java/appeng/client/gui/implementations/GuiFormationPlane.java
@@ -17,7 +17,10 @@ import org.lwjgl.input.Mouse;
 
 import appeng.api.config.FuzzyMode;
 import appeng.api.config.Settings;
+import appeng.api.config.Upgrades;
 import appeng.api.config.YesNo;
+import appeng.api.storage.data.IAEStackType;
+import appeng.client.gui.slots.VirtualMEPhantomSlot;
 import appeng.client.gui.widgets.GuiImgButton;
 import appeng.client.gui.widgets.GuiTabButton;
 import appeng.container.implementations.ContainerFormationPlane;
@@ -27,21 +30,35 @@ import appeng.core.sync.GuiBridge;
 import appeng.core.sync.network.NetworkHandler;
 import appeng.core.sync.packets.PacketConfigButton;
 import appeng.core.sync.packets.PacketSwitchGuis;
-import appeng.parts.automation.PartFormationPlane;
+import appeng.parts.automation.PartBaseFormationPlane;
+import appeng.tile.inventory.IAEStackInventory;
 
 public class GuiFormationPlane extends GuiUpgradeable {
 
     private GuiTabButton priority;
     private GuiImgButton placeMode;
+    private VirtualMEPhantomSlot[] configSlots;
+    protected final ContainerFormationPlane cfp;
 
-    public GuiFormationPlane(final InventoryPlayer inventoryPlayer, final PartFormationPlane te) {
+    public GuiFormationPlane(final InventoryPlayer inventoryPlayer, final PartBaseFormationPlane te) {
         super(new ContainerFormationPlane(inventoryPlayer, te));
         this.ySize = 251;
+        this.cfp = (ContainerFormationPlane) this.cvb;
+    }
+
+    @Override
+    public void initGui() {
+        super.initGui();
+        initVirtualSlots();
     }
 
     @Override
     protected void addButtons() {
-        this.placeMode = new GuiImgButton(this.guiLeft - 18, this.guiTop + 28, Settings.PLACE_BLOCK, YesNo.YES);
+        if (this.cfp.supportItemDrop()) {
+            this.placeMode = new GuiImgButton(this.guiLeft - 18, this.guiTop + 28, Settings.PLACE_BLOCK, YesNo.YES);
+            this.buttonList.add(this.placeMode);
+        }
+
         this.fuzzyMode = new GuiImgButton(
                 this.guiLeft - 18,
                 this.guiTop + 48,
@@ -56,7 +73,6 @@ public class GuiFormationPlane extends GuiUpgradeable {
                         GuiText.Priority.getLocal(),
                         itemRender));
 
-        this.buttonList.add(this.placeMode);
         this.buttonList.add(this.fuzzyMode);
     }
 
@@ -78,7 +94,26 @@ public class GuiFormationPlane extends GuiUpgradeable {
         }
 
         if (this.placeMode != null) {
-            this.placeMode.set(((ContainerFormationPlane) this.cvb).getPlaceMode());
+            this.placeMode.set(this.cfp.getPlaceMode());
+        }
+
+        this.updateSlotVisibility();
+    }
+
+    @Override
+    public void drawBG(int offsetX, int offsetY, int mouseX, int mouseY) {
+        super.drawBG(offsetX, offsetY, mouseX, mouseY);
+
+        final int capacity = this.cvb.getUpgradeable().getInstalledUpgrades(Upgrades.CAPACITY);
+
+        for (int i = 0; i < 7; i++) {
+            if (i >= capacity + 2) {
+                // fadeout slots
+                this.drawTexturedModalRect(offsetX + 7, offsetY + 28 + (18 * i), 7, 46, 162, 18);
+            } else {
+                // normal slots
+                this.drawTexturedModalRect(offsetX + 7, offsetY + 28 + (18 * i), 7, 28, 162, 18);
+            }
         }
     }
 
@@ -98,5 +133,37 @@ public class GuiFormationPlane extends GuiUpgradeable {
         } else if (btn == this.placeMode) {
             NetworkHandler.instance.sendToServer(new PacketConfigButton(this.placeMode.getSetting(), backwards));
         }
+    }
+
+    private void initVirtualSlots() {
+        this.configSlots = new VirtualMEPhantomSlot[63];
+        final IAEStackInventory inputInv = this.cfp.getConfig();
+        final int xo = 8;
+        final int yo = -133;
+
+        for (int y = 0; y < 7; y++) {
+            for (int x = 0; x < 9; x++) {
+                VirtualMEPhantomSlot slot = new VirtualMEPhantomSlot(
+                        xo + x * 18,
+                        yo + y * 18 + 9 * 18,
+                        inputInv,
+                        x + y * 9,
+                        this::acceptType);
+                this.configSlots[x + y * 9] = slot;
+                this.registerVirtualSlots(slot);
+            }
+        }
+    }
+
+    protected void updateSlotVisibility() {
+        final int capacity = this.cfp.getUpgradeable().getInstalledUpgrades(Upgrades.CAPACITY);
+
+        for (VirtualMEPhantomSlot slot : this.configSlots) {
+            slot.setHidden(slot.getSlotIndex() >= (18 + (9 * capacity)));
+        }
+    }
+
+    private boolean acceptType(VirtualMEPhantomSlot slot, IAEStackType<?> type, int mouseButton) {
+        return type == this.cfp.getStackType();
     }
 }
