+++ src/main/java/appeng/parts/automation/PartAnnihilationPlane.java
@@ -14,7 +14,6 @@ import java.util.List;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
-import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.player.EntityPlayer;
@@ -36,39 +35,26 @@ import appeng.api.config.PowerMultiplier;
 import appeng.api.config.YesNo;
 import appeng.api.networking.IGridNode;
 import appeng.api.networking.energy.IEnergyGrid;
-import appeng.api.networking.events.MENetworkChannelsChanged;
-import appeng.api.networking.events.MENetworkEventSubscribe;
-import appeng.api.networking.events.MENetworkPowerStatusChange;
 import appeng.api.networking.security.BaseActionSource;
 import appeng.api.networking.security.MachineSource;
 import appeng.api.networking.storage.IStorageGrid;
 import appeng.api.networking.ticking.IGridTickable;
 import appeng.api.networking.ticking.TickRateModulation;
 import appeng.api.networking.ticking.TickingRequest;
-import appeng.api.parts.IPart;
-import appeng.api.parts.IPartCollisionHelper;
-import appeng.api.parts.IPartHost;
-import appeng.api.parts.IPartRenderHelper;
 import appeng.api.storage.data.IAEItemStack;
 import appeng.client.texture.CableBusTextures;
 import appeng.core.settings.TickRates;
 import appeng.core.sync.packets.PacketTransitionEffect;
 import appeng.hooks.TickHandler;
 import appeng.me.GridAccessException;
-import appeng.parts.PartBasicState;
 import appeng.server.ServerHelper;
 import appeng.util.IWorldCallable;
 import appeng.util.Platform;
 import appeng.util.item.AEItemStack;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
-public class PartAnnihilationPlane extends PartBasicState implements IGridTickable, IWorldCallable<TickRateModulation> {
+public class PartAnnihilationPlane extends PartBaseAnnihilationPlane
+        implements IGridTickable, IWorldCallable<TickRateModulation> {
 
-    private static final IIcon SIDE_ICON = CableBusTextures.PartPlaneSides.getIcon();
-    private static final IIcon BACK_ICON = CableBusTextures.PartTransitionPlaneBack.getIcon();
-    private static final IIcon STATUS_ICON = CableBusTextures.PartMonitorSidesStatus.getIcon();
-    private static final IIcon ACTIVE_ICON = CableBusTextures.BlockAnnihilationPlaneOn.getIcon();
     private static final int MAX_CACHE_TIME = 60;
 
     private final BaseActionSource mySrc = new MachineSource(this);
@@ -92,137 +78,6 @@ public class PartAnnihilationPlane extends PartBasicState implements IGridTickab
     public TickRateModulation call(final World world) throws Exception {
         this.breaking = false;
         return this.breakBlock(true);
-    }
-
-    @Override
-    public void getBoxes(final IPartCollisionHelper bch) {
-        int minX = 1;
-        int minY = 1;
-        int maxX = 15;
-        int maxY = 15;
-
-        final IPartHost host = this.getHost();
-        if (host != null) {
-            final TileEntity te = host.getTile();
-
-            final int x = te.xCoord;
-            final int y = te.yCoord;
-            final int z = te.zCoord;
-
-            final ForgeDirection e = bch.getWorldX();
-            final ForgeDirection u = bch.getWorldY();
-
-            if (this.isAnnihilationPlane(
-                    te.getWorldObj().getTileEntity(x - e.offsetX, y - e.offsetY, z - e.offsetZ),
-                    this.getSide())) {
-                minX = 0;
-            }
-
-            if (this.isAnnihilationPlane(
-                    te.getWorldObj().getTileEntity(x + e.offsetX, y + e.offsetY, z + e.offsetZ),
-                    this.getSide())) {
-                maxX = 16;
-            }
-
-            if (this.isAnnihilationPlane(
-                    te.getWorldObj().getTileEntity(x - u.offsetX, y - u.offsetY, z - u.offsetZ),
-                    this.getSide())) {
-                minY = 0;
-            }
-
-            if (this.isAnnihilationPlane(
-                    te.getWorldObj().getTileEntity(x + u.offsetX, y + u.offsetY, z + u.offsetZ),
-                    this.getSide())) {
-                maxY = 16;
-            }
-        }
-
-        bch.addBox(5, 5, 14, 11, 11, 15);
-        bch.addBox(minX, minY, 15, maxX, maxY, bch.isBBCollision() ? 15 : 16);
-    }
-
-    @Override
-    @SideOnly(Side.CLIENT)
-    public void renderInventory(final IPartRenderHelper rh, final RenderBlocks renderer) {
-        rh.setTexture(SIDE_ICON, SIDE_ICON, BACK_ICON, this.getItemStack().getIconIndex(), SIDE_ICON, SIDE_ICON);
-
-        rh.setBounds(1, 1, 15, 15, 15, 16);
-        rh.renderInventoryBox(renderer);
-
-        rh.setBounds(5, 5, 14, 11, 11, 15);
-        rh.renderInventoryBox(renderer);
-    }
-
-    @Override
-    @SideOnly(Side.CLIENT)
-    public void renderStatic(final int x, final int y, final int z, final IPartRenderHelper rh,
-            final RenderBlocks renderer) {
-        this.renderStaticWithIcon(x, y, z, rh, renderer, ACTIVE_ICON);
-    }
-
-    protected void renderStaticWithIcon(final int x, final int y, final int z, final IPartRenderHelper rh,
-            final RenderBlocks renderer, final IIcon activeIcon) {
-        int minX = 1;
-
-        final ForgeDirection e = rh.getWorldX();
-        final ForgeDirection u = rh.getWorldY();
-
-        final TileEntity te = this.getHost().getTile();
-
-        if (this.isAnnihilationPlane(
-                te.getWorldObj().getTileEntity(x - e.offsetX, y - e.offsetY, z - e.offsetZ),
-                this.getSide())) {
-            minX = 0;
-        }
-
-        int maxX = 15;
-        if (this.isAnnihilationPlane(
-                te.getWorldObj().getTileEntity(x + e.offsetX, y + e.offsetY, z + e.offsetZ),
-                this.getSide())) {
-            maxX = 16;
-        }
-
-        int minY = 1;
-        if (this.isAnnihilationPlane(
-                te.getWorldObj().getTileEntity(x - u.offsetX, y - u.offsetY, z - u.offsetZ),
-                this.getSide())) {
-            minY = 0;
-        }
-
-        int maxY = 15;
-        if (this.isAnnihilationPlane(
-                te.getWorldObj().getTileEntity(x + u.offsetX, y + u.offsetY, z + u.offsetZ),
-                this.getSide())) {
-            maxY = 16;
-        }
-
-        final boolean isActive = (this.getClientFlags() & (PartBasicState.POWERED_FLAG | PartBasicState.CHANNEL_FLAG))
-                == (PartBasicState.POWERED_FLAG | PartBasicState.CHANNEL_FLAG);
-
-        this.setRenderCache(rh.useSimplifiedRendering(x, y, z, this, this.getRenderCache()));
-        rh.setTexture(
-                SIDE_ICON,
-                SIDE_ICON,
-                BACK_ICON,
-                isActive ? activeIcon : this.getItemStack().getIconIndex(),
-                SIDE_ICON,
-                SIDE_ICON);
-
-        rh.setBounds(minX, minY, 15, maxX, maxY, 16);
-        rh.renderBlock(x, y, z, renderer);
-
-        rh.setTexture(
-                STATUS_ICON,
-                STATUS_ICON,
-                BACK_ICON,
-                isActive ? activeIcon : this.getItemStack().getIconIndex(),
-                STATUS_ICON,
-                STATUS_ICON);
-
-        rh.setBounds(5, 5, 14, 11, 11, 15);
-        rh.renderBlock(x, y, z, renderer);
-
-        this.renderLights(x, y, z, rh, renderer);
     }
 
     @Override
@@ -296,11 +151,6 @@ public class PartAnnihilationPlane extends PartBasicState implements IGridTickab
                 }
             }
         }
-    }
-
-    @Override
-    public int cableConnectionRenderTo() {
-        return 1;
     }
 
     /**
@@ -387,28 +237,6 @@ public class PartAnnihilationPlane extends PartBasicState implements IGridTickab
         overflowEntity.motionZ = 0;
 
         w.spawnEntityInWorld(overflowEntity);
-    }
-
-    protected boolean isAnnihilationPlane(final TileEntity blockTileEntity, final ForgeDirection side) {
-        if (blockTileEntity instanceof IPartHost) {
-            final IPart p = ((IPartHost) blockTileEntity).getPart(side);
-            return p != null && p.getClass() == this.getClass();
-        }
-        return false;
-    }
-
-    @Override
-    @MENetworkEventSubscribe
-    public void chanRender(final MENetworkChannelsChanged c) {
-        this.onNeighborChanged();
-        this.getHost().markForUpdate();
-    }
-
-    @Override
-    @MENetworkEventSubscribe
-    public void powerRender(final MENetworkPowerStatusChange c) {
-        this.onNeighborChanged();
-        this.getHost().markForUpdate();
     }
 
     private TickRateModulation breakBlock(final boolean modulate) {
@@ -584,5 +412,10 @@ public class PartAnnihilationPlane extends PartBasicState implements IGridTickab
             final IAEItemStack overflow = this.storeItemStack(snaggedItem);
             this.spawnOverflow(overflow);
         }
+    }
+
+    @Override
+    public IIcon getActiveIcon() {
+        return CableBusTextures.BlockAnnihilationPlaneOn.getIcon();
     }
 }
