+++ src/main/java/appeng/parts/automation/PartFormationPlane.java
@@ -17,21 +17,18 @@ import java.util.List;
 
 import javax.annotation.Nonnull;
 
-import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Blocks;
-import net.minecraft.inventory.IInventory;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemFirework;
 import net.minecraft.item.ItemReed;
 import net.minecraft.item.ItemSkull;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.Vec3;
+import net.minecraft.util.IIcon;
 import net.minecraft.world.World;
 import net.minecraft.world.WorldServer;
 import net.minecraft.world.chunk.Chunk;
@@ -52,14 +49,7 @@ import appeng.api.config.Settings;
 import appeng.api.config.Upgrades;
 import appeng.api.config.YesNo;
 import appeng.api.networking.events.MENetworkCellArrayUpdate;
-import appeng.api.networking.events.MENetworkChannelsChanged;
-import appeng.api.networking.events.MENetworkEventSubscribe;
-import appeng.api.networking.events.MENetworkPowerStatusChange;
 import appeng.api.networking.security.BaseActionSource;
-import appeng.api.parts.IPart;
-import appeng.api.parts.IPartCollisionHelper;
-import appeng.api.parts.IPartHost;
-import appeng.api.parts.IPartRenderHelper;
 import appeng.api.storage.ICellContainer;
 import appeng.api.storage.IMEInventory;
 import appeng.api.storage.IMEInventoryHandler;
@@ -67,38 +57,22 @@ import appeng.api.storage.StorageChannel;
 import appeng.api.storage.data.IAEItemStack;
 import appeng.api.storage.data.IAEStackType;
 import appeng.api.storage.data.IItemList;
-import appeng.api.util.IConfigManager;
 import appeng.client.texture.CableBusTextures;
 import appeng.core.AEConfig;
-import appeng.core.sync.GuiBridge;
-import appeng.helpers.IPrimaryGuiIconProvider;
-import appeng.helpers.IPriorityHost;
 import appeng.me.GridAccessException;
 import appeng.me.storage.MEInventoryHandler;
-import appeng.parts.PartBasicState;
-import appeng.tile.inventory.AppEngInternalAEInventory;
-import appeng.tile.inventory.InvOperation;
 import appeng.util.Platform;
 import appeng.util.prioitylist.FuzzyPriorityList;
 import appeng.util.prioitylist.PrecisePriorityList;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
-public class PartFormationPlane extends PartUpgradeable
-        implements ICellContainer, IPriorityHost, IMEInventory<IAEItemStack>, IPrimaryGuiIconProvider {
+public class PartFormationPlane extends PartBaseFormationPlane implements ICellContainer, IMEInventory<IAEItemStack> {
 
     private final MEInventoryHandler myHandler = new MEInventoryHandler(this, ITEM_STACK_TYPE);
-    private final AppEngInternalAEInventory Config = new AppEngInternalAEInventory(this, 63);
     private EntityPlayer owner = null;
-    private int priority = 0;
-    private boolean wasActive = false;
-    private boolean blocked = false;
 
     public PartFormationPlane(final ItemStack is) {
         super(is);
 
-        this.getConfigManager().registerSetting(Settings.FUZZY_MODE, FuzzyMode.IGNORE_ALL);
-        this.getConfigManager().registerSetting(Settings.PLACE_BLOCK, YesNo.YES);
         this.updateHandler();
     }
 
@@ -108,7 +82,7 @@ public class PartFormationPlane extends PartUpgradeable
         this.owner = player;
     }
 
-    private void updateHandler() {
+    protected void updateHandler() {
         this.myHandler.setBaseAccess(AccessRestriction.WRITE);
         this.myHandler.setWhitelist(
                 this.getInstalledUpgrades(Upgrades.INVERTER) > 0 ? IncludeExclude.BLACKLIST : IncludeExclude.WHITELIST);
@@ -118,10 +92,8 @@ public class PartFormationPlane extends PartUpgradeable
 
         final int slotsToUse = 18 + this.getInstalledUpgrades(Upgrades.CAPACITY) * 9;
         for (int x = 0; x < this.Config.getSizeInventory() && x < slotsToUse; x++) {
-            final IAEItemStack is = this.Config.getAEStackInSlot(x);
-            if (is != null) {
-                priorityList.add(is);
-            }
+            if (!(this.Config.getAEStackInSlot(x) instanceof IAEItemStack ais)) continue;
+            priorityList.add(ais);
         }
 
         if (this.getInstalledUpgrades(Upgrades.FUZZY) > 0) {
@@ -141,249 +113,6 @@ public class PartFormationPlane extends PartUpgradeable
     }
 
     @Override
-    protected int getUpgradeSlots() {
-        return 5;
-    }
-
-    @Override
-    public void updateSetting(final IConfigManager manager, final Enum settingName, final Enum newValue) {
-        this.updateHandler();
-        this.getHost().markForSave();
-    }
-
-    @Override
-    public void onChangeInventory(final IInventory inv, final int slot, final InvOperation mc,
-            final ItemStack removedStack, final ItemStack newStack) {
-        super.onChangeInventory(inv, slot, mc, removedStack, newStack);
-
-        if (inv == this.Config) {
-            this.updateHandler();
-        }
-    }
-
-    @Override
-    public void upgradesChanged() {
-        this.updateHandler();
-    }
-
-    @Override
-    public void readFromNBT(final NBTTagCompound data) {
-        super.readFromNBT(data);
-        this.Config.readFromNBT(data, "config");
-        this.priority = data.getInteger("priority");
-        this.updateHandler();
-    }
-
-    @Override
-    public void writeToNBT(final NBTTagCompound data) {
-        super.writeToNBT(data);
-        this.Config.writeToNBT(data, "config");
-        data.setInteger("priority", this.priority);
-    }
-
-    @Override
-    public IInventory getInventoryByName(final String name) {
-        if (name.equals("config")) {
-            return this.Config;
-        }
-
-        return super.getInventoryByName(name);
-    }
-
-    @Override
-    @MENetworkEventSubscribe
-    public void powerRender(final MENetworkPowerStatusChange c) {
-        final boolean currentActive = this.getProxy().isActive();
-        if (this.wasActive != currentActive) {
-            this.wasActive = currentActive;
-            this.updateHandler(); // proxy.getGrid().postEvent( new MENetworkCellArrayUpdate() );
-            this.getHost().markForUpdate();
-        }
-    }
-
-    @MENetworkEventSubscribe
-    public void updateChannels(final MENetworkChannelsChanged changedChannels) {
-        final boolean currentActive = this.getProxy().isActive();
-        if (this.wasActive != currentActive) {
-            this.wasActive = currentActive;
-            this.updateHandler(); // proxy.getGrid().postEvent( new MENetworkCellArrayUpdate() );
-            this.getHost().markForUpdate();
-        }
-    }
-
-    @Override
-    public void getBoxes(final IPartCollisionHelper bch) {
-        int minX = 1;
-        int minY = 1;
-        int maxX = 15;
-        int maxY = 15;
-
-        final IPartHost host = this.getHost();
-        if (host != null) {
-            final TileEntity te = host.getTile();
-
-            final int x = te.xCoord;
-            final int y = te.yCoord;
-            final int z = te.zCoord;
-
-            final ForgeDirection e = bch.getWorldX();
-            final ForgeDirection u = bch.getWorldY();
-
-            if (this.isTransitionPlane(
-                    te.getWorldObj().getTileEntity(x - e.offsetX, y - e.offsetY, z - e.offsetZ),
-                    this.getSide())) {
-                minX = 0;
-            }
-
-            if (this.isTransitionPlane(
-                    te.getWorldObj().getTileEntity(x + e.offsetX, y + e.offsetY, z + e.offsetZ),
-                    this.getSide())) {
-                maxX = 16;
-            }
-
-            if (this.isTransitionPlane(
-                    te.getWorldObj().getTileEntity(x - u.offsetX, y - u.offsetY, z - u.offsetZ),
-                    this.getSide())) {
-                minY = 0;
-            }
-
-            if (this.isTransitionPlane(
-                    te.getWorldObj().getTileEntity(x + u.offsetX, y + u.offsetY, z + u.offsetZ),
-                    this.getSide())) {
-                maxY = 16;
-            }
-        }
-
-        bch.addBox(5, 5, 14, 11, 11, 15);
-        bch.addBox(minX, minY, 15, maxX, maxY, 16);
-    }
-
-    @Override
-    @SideOnly(Side.CLIENT)
-    public void renderInventory(final IPartRenderHelper rh, final RenderBlocks renderer) {
-        rh.setTexture(
-                CableBusTextures.PartPlaneSides.getIcon(),
-                CableBusTextures.PartPlaneSides.getIcon(),
-                CableBusTextures.PartTransitionPlaneBack.getIcon(),
-                this.getItemStack().getIconIndex(),
-                CableBusTextures.PartPlaneSides.getIcon(),
-                CableBusTextures.PartPlaneSides.getIcon());
-
-        rh.setBounds(1, 1, 15, 15, 15, 16);
-        rh.renderInventoryBox(renderer);
-
-        rh.setBounds(5, 5, 14, 11, 11, 15);
-        rh.renderInventoryBox(renderer);
-    }
-
-    @Override
-    @SideOnly(Side.CLIENT)
-    public void renderStatic(final int x, final int y, final int z, final IPartRenderHelper rh,
-            final RenderBlocks renderer) {
-        int minX = 1;
-
-        final ForgeDirection e = rh.getWorldX();
-        final ForgeDirection u = rh.getWorldY();
-
-        final TileEntity te = this.getHost().getTile();
-
-        if (this.isTransitionPlane(
-                te.getWorldObj().getTileEntity(x - e.offsetX, y - e.offsetY, z - e.offsetZ),
-                this.getSide())) {
-            minX = 0;
-        }
-
-        int maxX = 15;
-        if (this.isTransitionPlane(
-                te.getWorldObj().getTileEntity(x + e.offsetX, y + e.offsetY, z + e.offsetZ),
-                this.getSide())) {
-            maxX = 16;
-        }
-
-        int minY = 1;
-        if (this.isTransitionPlane(
-                te.getWorldObj().getTileEntity(x - u.offsetX, y - u.offsetY, z - u.offsetZ),
-                this.getSide())) {
-            minY = 0;
-        }
-
-        int maxY = 15;
-        if (this.isTransitionPlane(
-                te.getWorldObj().getTileEntity(x + u.offsetX, y + u.offsetY, z + u.offsetZ),
-                this.getSide())) {
-            maxY = 16;
-        }
-
-        final boolean isActive = (this.getClientFlags() & (PartBasicState.POWERED_FLAG | PartBasicState.CHANNEL_FLAG))
-                == (PartBasicState.POWERED_FLAG | PartBasicState.CHANNEL_FLAG);
-
-        this.setRenderCache(rh.useSimplifiedRendering(x, y, z, this, this.getRenderCache()));
-        rh.setTexture(
-                CableBusTextures.PartPlaneSides.getIcon(),
-                CableBusTextures.PartPlaneSides.getIcon(),
-                CableBusTextures.PartTransitionPlaneBack.getIcon(),
-                isActive ? CableBusTextures.BlockFormPlaneOn.getIcon() : this.getItemStack().getIconIndex(),
-                CableBusTextures.PartPlaneSides.getIcon(),
-                CableBusTextures.PartPlaneSides.getIcon());
-
-        rh.setBounds(minX, minY, 15, maxX, maxY, 16);
-        rh.renderBlock(x, y, z, renderer);
-
-        rh.setTexture(
-                CableBusTextures.PartMonitorSidesStatus.getIcon(),
-                CableBusTextures.PartMonitorSidesStatus.getIcon(),
-                CableBusTextures.PartTransitionPlaneBack.getIcon(),
-                isActive ? CableBusTextures.BlockFormPlaneOn.getIcon() : this.getItemStack().getIconIndex(),
-                CableBusTextures.PartMonitorSidesStatus.getIcon(),
-                CableBusTextures.PartMonitorSidesStatus.getIcon());
-
-        rh.setBounds(5, 5, 14, 11, 11, 15);
-        rh.renderBlock(x, y, z, renderer);
-
-        this.renderLights(x, y, z, rh, renderer);
-    }
-
-    @Override
-    public void onNeighborChanged() {
-        final TileEntity te = this.getHost().getTile();
-        final World w = te.getWorldObj();
-        final ForgeDirection side = this.getSide();
-
-        final int x = te.xCoord + side.offsetX;
-        final int y = te.yCoord + side.offsetY;
-        final int z = te.zCoord + side.offsetZ;
-
-        this.blocked = !w.getBlock(x, y, z).isReplaceable(w, x, y, z);
-    }
-
-    @Override
-    public int cableConnectionRenderTo() {
-        return 1;
-    }
-
-    @Override
-    public boolean onPartActivate(final EntityPlayer player, final Vec3 pos) {
-        if (!player.isSneaking()) {
-            if (Platform.isClient()) {
-                return true;
-            }
-
-            Platform.openGUI(player, this.getHost().getTile(), this.getSide(), GuiBridge.GUI_FORMATION_PLANE);
-            return true;
-        }
-
-        return false;
-    }
-
-    private boolean isTransitionPlane(final TileEntity blockTileEntity, final ForgeDirection side) {
-        if (blockTileEntity instanceof IPartHost) {
-            final IPart p = ((IPartHost) blockTileEntity).getPart(side);
-            return p instanceof PartFormationPlane;
-        }
-        return false;
-    }
-
-    @Override
     @NotNull
     public List<IMEInventoryHandler> getCellArray(final IAEStackType<?> type) {
         if (this.getProxy().isActive() && type == ITEM_STACK_TYPE) {
@@ -392,18 +121,6 @@ public class PartFormationPlane extends PartUpgradeable
             return Handler;
         }
         return new ArrayList<>();
-    }
-
-    @Override
-    public int getPriority() {
-        return this.priority;
-    }
-
-    @Override
-    public void setPriority(final int newValue) {
-        this.priority = newValue;
-        this.getHost().markForSave();
-        this.updateHandler();
     }
 
     @Override
@@ -655,7 +372,17 @@ public class PartFormationPlane extends PartUpgradeable
     }
 
     @Override
-    public ItemStack getPrimaryGuiIcon() {
-        return AEApi.instance().definitions().parts().formationPlane().maybeStack(1).orNull();
+    public @NotNull IAEStackType<?> getStackType() {
+        return ITEM_STACK_TYPE;
+    }
+
+    @Override
+    public boolean supportItemDrop() {
+        return true;
+    }
+
+    @Override
+    public IIcon getActiveIcon() {
+        return CableBusTextures.BlockFormPlaneOn.getIcon();
     }
 }
